Tweaked commands
- 'balance' command renamed to 'card'
- 'card' command tweaked
- disabled a few other non-functional commands
