+++ bot/commands/!test.js
